# WIP: refactor graph in LazyBridgeOptimizer