The quest to #1347. My goal is a `graph::Union{Nothing,Graph}`, that we postpone initializing for as long as possible.

## Before 

```Julia
Running: clp 
 10.167812 seconds (25.10 M allocations: 1.464 GiB, 6.86% gc time, 22.51% compilation time)
  0.000808 seconds (1.58 k allocations: 114.773 KiB)
```

## After 

```Julia
Running: clp 
 10.115014 seconds (24.83 M allocations: 1.445 GiB, 6.91% gc time, 21.48% compilation time)
  0.001068 seconds (1.53 k allocations: 109.242 KiB)
```