@odow 2021-05-20
Closing because this didn't really work. We have a better plan at the JuMP-side.